requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.2.10
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
django-heroku==0.3.1
django-import-export==2.5.0
django-sms==0.5.0
djangorestframework==3.12.2
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
kombu==5.1.0
librabbitmq==2.0.0
MarkupPy==1.14
MouseInfo==0.1.3
odfpy==1.4.1
openpyxl==3.0.7
pika==1.2.0
Pillow==8.3.1
platformdirs==2.2.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.2
python-decouple==3.4
PyTweening==1.0.3
pytz==2021.1
pywhatkit==5.1
PyYAML==5.4.1
requests==2.26.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
unicorn==1.0.3
urllib3==1.26.6
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
whitenoise==5.3.0
wikipedia==1.4.0
xlrd==2.0.1
xlwt==1.3.0
